Changed chambers field to a double rather than an unsignled long int.

+++ src/cls/setup/sim.hpp
@@ -49,8 +49,8 @@ namespace arc
             const unsigned long int m_num_phot; //! Number of photons to simulate.
 
             //  -- Roulette --
-            const double            m_roulette_weight;      //! Roulette threshold.
-            const unsigned long int m_roulette_chambers;    //! Number of roulette chambers.
+            const double m_roulette_weight;     //! Roulette threshold.
+            const double m_roulette_chambers;   //! Number of roulette chambers.
 
             //  -- Equipment --
             const phys::Material             m_aether;  //! Aether material.
